requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
colorlog==6.9.0
frozenlist==1.5.0
hikari==2.2.0
idna==3.10
multidict==6.2.0
propcache==0.3.1
python-dotenv==1.1.0
yarl==1.18.3
